refactor: move the group anchor helper out of the component module
