+++ src/best-practices/anchors.ts
@@ -1,0 +1,9 @@
+// -- Anchors -------------------------------------------------------------------
+
+function groupAnchorId(groupId: string): string {
+  return `best-practice-${groupId}`;
+}
+
+// -- Exports -------------------------------------------------------------------
+
+export { groupAnchorId };
