+++ src/best-practices/rule-list.browser.test.tsx
@@ -1,6 +1,7 @@
 import { describe, expect, it } from "vitest";
+import { groupAnchorId } from "@/best-practices/anchors";
 import { BEST_PRACTICE_GROUPS } from "@/best-practices/groups";
-import { groupAnchorId, RuleList } from "@/best-practices/rule-list";
+import { RuleList } from "@/best-practices/rule-list";
 import { render } from "@/test/render";
 
 // -- Helpers -------------------------------------------------------------------
