+++ src/ui/help-sections/best-practices.tsx
@@ -1,7 +1,8 @@
 import { useReducedMotion } from "motion/react";
+import { groupAnchorId } from "@/best-practices/anchors";
 import { BEST_PRACTICE_GROUPS } from "@/best-practices/groups";
 import type { RuleGroup } from "@/best-practices/model";
-import { groupAnchorId, RuleList } from "@/best-practices/rule-list";
+import { RuleList } from "@/best-practices/rule-list";
 import { PROSE } from "@/ui/typography";
 
 // -- Interfaces ---------------------------------------------------------------
